requirements.txt
streamlit
requests
Pillow
openai
python-dotenv